==> 01_weather_concat.ipynb <==
## Concatenating weather data

import sys
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np
import pandas as pd
from datetime import datetime, date
import time
import matplotlib.pyplot as plt

### Concatenate all weather data

df_2015 = pd.read_csv('data/raw/weather/weather_2015.csv') 
df_2016 = pd.read_csv('data/raw/weather/weather_2016.csv') 
df_2017 = pd.read_csv('data/raw/weather/weather_2017.csv') 
df_2018 = pd.read_csv('data/raw/weather/weather_2018.csv') 
df_2019 = pd.read_csv('data/raw/weather/weather_2019.csv') 

frames_all = [df_2015, df_2016, df_2017, df_2018, df_2019]
result_all = pd.concat(frames_all)
result_all = result_all.drop(columns=["Unnamed: 0"])

result_all.tail()

result_all.to_csv('data/interim/weather/weather_all.csv')

## Process Weather Data

#### Get average weather infomation per day

result_all = result_all.drop(columns=["Time","Wind","Condition"])
result_all["DATE"] = result_all["date"].apply(lambda x: x.split(" ")[0])
result_all = result_all.drop(columns=["date"])

result_all_by_day = result_all.groupby("DATE").mean()
result_all_by_day = result_all_by_day.reset_index()
result_all_by_day.to_csv('data/interim/weather/weather_all_by_day.csv')

#### Get average weather infomation per hour group
- divide a day into 3 periods : 12AM-8AM, 8AM-4PM, 4PM-12AM
- average weather infomation within the period

df_weather_all = pd.read_csv('data/interim/weather/weather_all.csv')
df_weather_all = df_weather_all.drop(columns=["Unnamed: 0"])
df_weather_all["DATE"] = df_weather_all["date"].apply(lambda x: x.split(" ")[0])
df_weather_all["HOUR"] = df_weather_all["date"].apply(lambda x: datetime.strptime(x, '%Y-%m-%d %H:%M:%S').hour)

def what_hourgroup(hr):
    if hr >= 0 and hr <=7:
        return 1
    if hr >= 8 and hr <=15:
        return 2
    if hr >= 16 and hr <=23:
        return 3

df_weather_all["Hour_Grp"] = df_weather_all["HOUR"].apply(lambda x: what_hourgroup(x))

df_weather_all = df_weather_all.drop(columns=["Time","Wind","Condition","HOUR","date"])
#df_weather_all = df_weather_all.drop(columns=["date"])

df_weather_all=df_weather_all.groupby(["DATE","Hour_Grp"]).mean().reset_index()
df_weather_all.to_csv('data/interim/weather/weather_all_by_day_hourgroup.csv')
