==> 04_data_prep_merge_with_weather.ipynb <==
import sys
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np
import pandas as pd
from datetime import datetime, date
import time
import matplotlib.pyplot as plt

### Read In Crime data table

df_train = pd.read_csv('./data/interim/crime_data/fullTrainTestCrimeData_train_police_rand_2019_06_27.csv') 
df_valid = pd.read_csv('./data/interim/crime_data/fullTrainTestCrimeData_valid_police_rand_2019_06_27.csv') 
df_tests = pd.read_csv('./data/interim/crime_data/fullTrainTestCrimeData_test_police_rand_2019_06_27.csv')

df_valid.head(5)

df_train = df_train.drop(columns=["Unnamed: 0"])
df_valid = df_valid.drop(columns=["Unnamed: 0"])
df_tests = df_tests.drop(columns=["Unnamed: 0"])

### Create the column DAY_OF_WEEK

map_day = {'Monday':1, 'Tuesday':2,'Wednesday':3, 'Thursday':4, 'Friday':5, 
           'Saturday':6, 'Sunday':7}
df_train["DAY_OF_WEEK_NUM"] = df_train["DAY_OF_WEEK"].apply(lambda x: map_day[x])
df_valid["DAY_OF_WEEK_NUM"] = df_valid["DAY_OF_WEEK"].apply(lambda x: map_day[x])
df_tests["DAY_OF_WEEK_NUM"] = df_tests["DAY_OF_WEEK"].apply(lambda x: map_day[x])

### Read In the Weather info

df_weather_all_by_day = pd.read_csv('./data/interim/weather/weather_all_by_day_hourgroup.csv')
df_weather_all_by_day = df_weather_all_by_day.drop(columns=["Unnamed: 0"])
print(df_weather_all_by_day.shape)

#### More cleaning of weather table

if df_weather_all_by_day[df_weather_all_by_day["DATE"] == '2018-08-26'].shape[0] <3:
    new_series = df_weather_all_by_day[df_weather_all_by_day["DATE"] == '2018-08-26'].iloc[0]
    new_series["Hour_Grp"] = 1
    df_weather_all_by_day = df_weather_all_by_day.append(new_series , ignore_index=True)
df_weather_all_by_day[df_weather_all_by_day["DATE"] == '2018-08-26']

print (df_weather_all_by_day.shape)

### Mearge Weather with Crime

merged_df_train = pd.merge(df_train, df_weather_all_by_day,  how='left', left_on=['DATE','Hour_Grp'], right_on = ['DATE','Hour_Grp'])
merged_df_valid = pd.merge(df_valid, df_weather_all_by_day,  how='left', left_on=['DATE','Hour_Grp'], right_on = ['DATE','Hour_Grp'])
merged_df_tests = pd.merge(df_tests, df_weather_all_by_day,  how='left', left_on=['DATE','Hour_Grp'], right_on = ['DATE','Hour_Grp'])
print (merged_df_train.shape)
print (merged_df_valid.shape)
print (merged_df_tests.shape)

merged_df_train.to_csv('./data/processed/crime_data/fullTrainTestCrimeData_train_weather_by_hourgroup_police_rand_2019_06_27.csv')
merged_df_valid.to_csv('./data/processed/crime_data/fullTrainTestCrimeData_valid_weather_by_hourgroup_police_rand_2019_06_27.csv')
merged_df_tests.to_csv('./data/processed/crime_data/fullTrainTestCrimeData_tests_weather_by_hourgroup_police_rand_2019_06_27.csv')
